Fix compile warnings

+++ src/platform/file_manager.c
@@ -317,9 +317,9 @@ int platform_file_manager_list_directory_contents(
                 // Skip current (.), parent (..) and hidden directories (.*)
                 continue;
             }
-            match = callback(name, file_info.st_mtime);
+            match = callback(name, (long) file_info.st_mtime);
         } else if (file_has_extension(name, extension)) {
-            match = callback(name, file_info.st_mtime);
+            match = callback(name, (long) file_info.st_mtime);
         }
         if (match == LIST_MATCH) {
             break;
